requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anthropic==0.3.11
anyio==3.7.1
asn1crypto==1.4.0
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
base58==2.1.1
bcrypt==4.1.2
beautifulsoup4==4.12.2
bech32==1.2.0
bip-utils==2.8.0
bitarray==2.9.2
blinker==1.7.0
blis==0.7.11
blspy==2.0.2
cached-property==1.5.2
cachetools==5.3.2
catalogue==2.0.10
cbor2==5.6.1
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==2.1.1
chroma-hnswlib==0.7.3
chromadb==0.4.13
click==8.0.2
cloudpathlib==0.16.0
coincurve==18.0.0
coloredlogs==15.0.1
confection==0.1.4
coverage==7.4.1
crcmod==1.7
cryptography==42.0.2
cssselect==1.2.0
cymem==2.0.8
cytoolz==0.12.3
dataclasses-json==0.6.4
distlib==0.3.8
distro==1.9.0
docker==6.1.2
dockerpty==0.4.1
docopt==0.6.2
docstring-parser==0.15
ecdsa==0.16.1
ed25519-blake2b==1.4.1
eth-abi==4.0.0
eth-account==0.8.0
eth-hash==0.6.0
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.5.2
eth-utils==2.2.0
exceptiongroup==1.2.0
faiss-cpu==1.7.4
fastapi==0.109.2
filelock==3.13.1
Flask==2.1.3
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
google-api-core==2.17.1
google-api-python-client==2.95.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
googlesearch-python==1.2.3
gql==3.5.0
graphql-core==3.2.3
grpcio==1.53.0
h11==0.14.0
hexbytes==0.3.1
httpcore==1.0.3
httplib2==0.22.0
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.20.3
humanfriendly==10.0
hypothesis==6.21.6
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
iniconfig==2.0.0
ipfshttpclient==0.8.0a2
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.3.3
langcodes==3.3.0
langsmith==0.0.92
lru-dict==1.2.0
lxml==5.1.0
markdown-it-py==3.0.0
markdownify==0.11.6
MarkupSafe==2.1.5
marshmallow==3.20.2
mdurl==0.1.2
mech-client==0.2.9
monotonic==1.6
morphys==1.0
mpmath==1.3.0
multiaddr==0.0.9
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
netaddr==1.1.0
numexpr==2.9.0
numpy==1.26.4
onnxruntime==1.17.0
open-aea==1.42.0
open-aea-cli-ipfs==1.42.0
open-aea-cosmpy==0.6.7
open-aea-ledger-cosmos==1.42.0
open-aea-ledger-ethereum==1.42.0
open-aea-test-autonomy==0.13.8
open-autonomy==0.13.8
openai==1.11.0
overrides==7.7.0
packaging==23.2
pandas==2.2.1
paramiko==3.4.0
parsimonious==0.9.0
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
posthog==3.4.1
preshed==3.0.9
protobuf==4.24.4
pulsar-client==3.4.0
py==1.11.0
py-ecc==6.0.0
py-multibase==1.0.3
py-multicodec==0.2.1
py-sr25519-bindings==0.2.0
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==2.6.1
pydantic_core==2.16.2
pydeck==0.8.1b0
Pygments==2.17.2
pymultihash==0.8.2
PyNaCl==1.5.0
pyparsing==3.1.1
PyPDF2==3.0.1
PyPika==0.48.9
pyrsistent==0.20.0
pytest==7.2.1
pytest-asyncio==0.20.3
pytest-cov==4.0.0
pytest-randomly==3.12.0
pytest-rerunfailures==11.0
python-baseconv==1.2.2
python-dateutil==2.8.2
python-dotenv==0.17.1
pytz==2022.2.1
pyunormalize==15.1.0
PyYAML==6.0.1
readability-lxml==0.8.1
regex==2023.12.25
requests==2.28.1
rich==13.7.1
rlp==3.0.0
rsa==4.9
scikit-learn==1.3.1
scipy==1.12.0
semver==2.13.0
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.27
srsly==2.4.8
starlette==0.36.3
streamlit==1.31.1
sympy==1.12
tavily-python==0.3.0
tenacity==8.2.3
texttable==1.6.7
thinc==8.2.3
threadpoolctl==3.3.0
tiktoken==0.5.1
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
tomte==0.2.15
toolz==0.12.1
tornado==6.4
tox==3.28.0
tqdm==4.66.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.26.18
uvicorn==0.27.1
uvloop==0.19.0
validators==0.22.0
valory-docker-compose==1.29.3
varint==1.0.2
virtualenv==20.25.0
wasabi==1.1.2
watchdog==4.0.0
watchfiles==0.21.0
weasel==0.3.4
web3==6.15.1
websocket-client==0.59.0
websockets==12.0
Werkzeug==2.0.3
yarl==1.9.4
zipp==3.17.0
